+++ src/new-components/base-button/base-button.tsx
@@ -47,6 +47,10 @@ type CommonProps = {
      * A tooltip linked to the button element.
      */
     tooltip?: TooltipProps['content']
+    /**
+     * The distance between the button element and the linked tooltip.
+     */
+    tooltipGapSize?: TooltipProps['gapSize']
 }
 
 type IconButtonProps = {
@@ -84,6 +88,7 @@ export const BaseButton = polymorphicComponent<'div', BaseButtonProps>(function 
         disabled = false,
         loading = false,
         tooltip,
+        tooltipGapSize,
         onClick,
         exceptionallySetClassName,
         children,
@@ -135,7 +140,9 @@ export const BaseButton = polymorphicComponent<'div', BaseButtonProps>(function 
     // If it's an icon-only button, make sure it uses the aria-label as tooltip if no tooltip was provided
     const tooltipContent = icon ? tooltip ?? props['aria-label'] : tooltip
     return tooltipContent ? (
-        <Tooltip content={tooltipContent}>{buttonElement}</Tooltip>
+        <Tooltip content={tooltipContent} gapSize={tooltipGapSize}>
+            {buttonElement}
+        </Tooltip>
     ) : (
         buttonElement
     )